requirements.txt
pandas
psycopg2-binary
tqdm
SQLAlchemy
python-dateutil